+++ frontend-shared/src/browser-compatibility-utils.js
