+++ frontend-shared/src/index.js
@@ -1,3 +1,8 @@
-module.exports = {
-  ...require('./svg-icon'),
-};
+// Components
+export { SvgIcon, registerIcons } from './svg-icon';
+
+// Hooks
+export { useElementShouldClose } from './hooks/use-element-should-close';
+
+// Utilities
+export { normalizeKeyName } from './browser-compatibility-utils';
