+++ frontend-shared/src/test/browser-compatiblity-utils-test.js
