+++ src/sidebar/components/MenuKeyboardNavigation.js
@@ -1,8 +1,7 @@
+import { normalizeKeyName } from '@hypothesis/frontend-shared';
 import { createElement } from 'preact';
 import { useEffect, useRef } from 'preact/hooks';
 import propTypes from 'prop-types';
-
-import { normalizeKeyName } from '../../shared/browser-compatibility-utils';
 
 function isElementVisible(element) {
   return element.offsetParent !== null;
